pyproject.toml
[project]
name = "learn-hello-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tavily-python>=0.7.23",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
